Refactor AIApp to load API key directly from environment variable and remove redundant property; update tests to use new mock environment variables
